EDSC-4327: Updates node, sass, bootstrap, and improves demo page.
