+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,37 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+Before contributing to this project, please review [Contribution.md](https://github.com/nasa/edsc-echoforms/blob/master/CONTRIBUTING.md).
+
+### Description
+A clear and concise description of what the bug is.
+
+**Reproduction Steps**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**System Information (please complete the following information):**
+ - OS: [e.g. iOS]
+ - Browser [e.g. chrome, safari]
+ - Version [e.g. 22]
+
+### Additional context
+Add any other context about the problem here.
+
+### Acceptance Criteria
+Steps to produce expected behavior.
