+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,0 +1,18 @@
+---
+name: Custom issue template
+about: Template for general tasks and improvements
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+Please review [Contribution.md](https://github.com/nasa/edsc-echoforms/blob/master/CONTRIBUTING.md) before contributing to this project.
+
+### Description
+Describe the purpose of the issue
+
+### Files
+Attach any relevant images or files
+
+### Acceptance Criteria
