+++ .github/workflows/ci.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
         matrix:
-          node-version: ['lts/hydrogen']
+          node-version: ['lts/jod']
     steps:
     - name: Checkout code
       uses: actions/checkout@v4
@@ -40,7 +40,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        node-version: ['lts/hydrogen']
+        node-version: ['lts/iron', 'lts/jod']
     steps:
     - name: Checkout code
       uses: actions/checkout@v4
@@ -69,9 +69,15 @@ jobs:
       uses: codecov/codecov-action@v4
       env:
         CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+  tests-passed:
+    needs: [tests]
+    runs-on: ubuntu-latest
+    steps:
+    - name: All Tests Have Passed
+      run: 'echo true'
   deploy:
     if: success() && github.ref == 'refs/heads/main' # only run on main success
-    needs: [tests] # only run after jest and cypress jobs complete
+    needs: [tests-passed] # only run after all test jobs complete
     runs-on: ubuntu-latest
     steps:
     - name: Checkout code
@@ -91,7 +97,7 @@ jobs:
         git push ecc $GITHUB_SHA:refs/heads/main
   deploy-gh-pages:
     if: success() && github.ref == 'refs/heads/main' # only run on main success
-    needs: [tests] # only run after jest and cypress jobs complete
+    needs: [tests-passed] # only run after all test jobs complete
     runs-on: ubuntu-latest
     steps:
     - name: Checkout code
