+++ .nvmrc
@@ -1,1 +1,1 @@
-lts/hydrogen
+lts/jod
