+++ babel.config.js
@@ -3,23 +3,22 @@ module.exports = {
     [
       '@babel/preset-env', {
         targets: {
-          node: '8.10',
-          esmodules: true,
-          ie: '10'
+          node: '22',
+          esmodules: true
         }
       }
     ],
     '@babel/preset-react'
   ],
+  env: {
+    test_cypress: {
+      plugins: ['istanbul']
+    }
+  },
   sourceType: 'unambiguous',
   plugins: [
     '@babel/plugin-proposal-object-rest-spread',
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-transform-runtime'
-  ],
-  env: {
-    test: {
-      plugins: ['istanbul']
-    }
-  }
+  ]
 }
