+++ example/src/styles.css
@@ -16,7 +16,3 @@
   width: 100%;
   height: 10em;
 }
-
-body {
-  padding: 0 5em;
-}
