+++ src/css/variables/_index.scss
@@ -1,0 +1,1 @@
+@forward "vendor";
